search for encoder
