+++ web/server/entrypoint_arm64.go
@@ -3,26 +3,20 @@
 package server
 
 import (
-	"strings"
-
 	"go.viam.com/rdk/gostream"
 	"go.viam.com/rdk/gostream/codec/h264"
+	"go.viam.com/rdk/gostream/codec/h264/ffmpeg/avcodec"
 	"go.viam.com/rdk/gostream/codec/opus"
 	"go.viam.com/rdk/gostream/codec/x264"
-	"go.viam.com/rdk/utils"
 )
 
-var streamConfig gostream.StreamConfig
-
-func init() {
-	if osInfo, err := utils.DetectOSInformation(); err == nil && strings.Contains(osInfo.Device, "Raspberry Pi") {
+func makeStreamConfig() gostream.StreamConfig {
+	var streamConfig gostream.StreamConfig
+	if avcodec.FindEncoderByName(h264.H264_V4L2M2M) != nil {
 		streamConfig.VideoEncoderFactory = h264.NewEncoderFactory()
 	} else {
 		streamConfig.VideoEncoderFactory = x264.NewEncoderFactory()
 	}
-}
-
-func makeStreamConfig() gostream.StreamConfig {
 	streamConfig.AudioEncoderFactory = opus.NewEncoderFactory()
 	return streamConfig
 }
